You / their, count nouns, and definitive articles
It's complicated, but English describes essences by using the word "the," and/or by pluralizing the noun that conveys the essence. When someone does differently, it sounds non native.

You is okay to use when you're describing a narrative, but not for instructions, because doing so makes the instructions technically incorrect.

I stuck with the convention of capitalizing Time, but it feels strange that it's the only token that isn't in all caps. Oh well.